requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
confection==0.1.0
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
deap==1.4.1
Django==4.2.3
django-crontab==0.7.1
django-filter==23.2
djangorestframework==3.14.0
filelock==3.12.2
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.16.4
hyperframe==5.2.0
idna==2.10
imbalanced-learn==0.11.0
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
libretranslatepy==2.1.1
lightning-utilities==0.9.0
lit==16.0.6
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mysqlclient==2.2.0
networkx==3.1
nltk==3.8.1
numpy==1.25.2
# DONT INSTALL IF YOU DON'T HAVE AN NVIDIA GRAPHICS CARD, IT CAN CAUSE ISSUES
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cuda-cupti-cu11==11.7.101
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.2.10.91
# nvidia-cusolver-cu11==11.4.0.1
# nvidia-cusparse-cu11==11.7.4.91
# nvidia-nccl-cu11==2.14.3
# nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==10.0.0
preshed==3.0.8
pydantic==1.10.12
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-lightning==2.0.6
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==2.4.7
sympy==1.12
textblob==0.17.1
thinc==8.1.10
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchmetrics==1.0.1
tqdm==4.65.0
transformers==4.31.0
# Throws error on windows avoid it
# triton
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wasabi==1.1.2
yarl==1.9.2
pylint-django